requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.2.0
python-pptx==1.0.2
email-validator==2.2.0
openpyxl==3.1.2
langchain==0.3.26
langchain-community==0.3.27
langchain-text-splitters==0.3.8
langchain-google-genai==2.0.10
faiss-cpu==1.11.0.post1
google-generativeai==0.8.5
pydantic==2.11.7
requests==2.28.2
streamlit==1.47.0
